admin/requirements.txt
# Core AWS and environment dependencies
boto3==1.28.0
python-dotenv==1.0.0

# Hugging Face for model downloads
huggingface_hub

# FastAPI UI dependencies
fastapi==0.95.2
uvicorn[standard]==0.22.0
jinja2==3.1.2
python-multipart==0.0.6
requests==2.31.0
